feat(config): support derby do snapshot

+++ config/src/main/java/com/alibaba/nacos/config/server/service/PersistService.java
@@ -962,7 +962,7 @@ public class PersistService {
     public long findConfigMaxId() {
         String sql = "SELECT max(id) FROM config_info";
         try {
-            return jt.queryForObject(sql, Integer.class);
+            return jt.queryForObject(sql, Long.class);
         } catch (NullPointerException e) {
             return 0;
         }
@@ -1696,7 +1696,7 @@ public class PersistService {
         if (result == null) {
             throw new IllegalArgumentException("aggrConfigInfoCount error");
         }
-        return result.intValue();
+        return result;
     }
 
     public int aggrConfigInfoCountIn(String dataId, String group, String tenant, List<String> datumIds) {
@@ -1735,7 +1735,7 @@ public class PersistService {
         if (result == null) {
             throw new IllegalArgumentException("aggrConfigInfoCount error");
         }
-        return result.intValue();
+        return result;
     }
 
     /**
